[4주차] 전수현 / [feat] 댓글 CRUD 기능 구현

+++ src/main/java/com/blog/domain/posts/api/PostsRestController.java
@@ -1,6 +1,9 @@
 package com.blog.domain.posts.api;
 
 import com.blog.common.response.ApiResponse;
+import com.blog.common.response.CustomException;
+import com.blog.common.response.ErrorCode;
+import com.blog.common.response.ResponseCode;
 import com.blog.domain.posts.api.dto.request.PostsRequest;
 import com.blog.domain.posts.api.dto.response.PostListResponse;
 import com.blog.domain.posts.api.dto.response.PostResponse;
@@ -24,14 +27,18 @@ public class PostsRestController {
     // 생성
     @PostMapping()
     public ApiResponse<String> createPost(
-            @RequestHeader("Authorization") String authHeader,
+            @RequestHeader(value = "Authorization", required = false) String authorization,
             @RequestBody PostsRequest request){
 
-        int userId = tokenService.extractUserIdFromHeader(authHeader);
+        if (authorization == null || authorization.isEmpty()){
+            throw new CustomException(ErrorCode.INVALID_TOKEN);
+        }
+
+        int userId = tokenService.extractUserIdFromHeader(authorization);
 
         postsService.createPost(userId, request);
 
-        return ApiResponse.ok("게시글이 작성 되었습니다.");
+        return ApiResponse.ok(ResponseCode.POST_CREATE_SUCCESS);
     }
 
     // 목록 조회
@@ -53,25 +60,33 @@ public class PostsRestController {
     // 수정
     @PatchMapping("/{postId}")
     public ApiResponse<String> updatePost(
-            @RequestHeader("Authorization") String authHeader,
+            @RequestHeader(value = "Authorization", required = false) String authorization,
             @PathVariable("postId") int postId,
             @RequestBody PostsRequest request){
 
-        int userId = tokenService.extractUserIdFromHeader(authHeader);
+        if (authorization == null || authorization.isEmpty()){
+            throw new CustomException(ErrorCode.INVALID_TOKEN);
+        }
+
+        int userId = tokenService.extractUserIdFromHeader(authorization);
         postsService.updatePost(userId, postId, request);
 
-        return ApiResponse.ok("수정 성공 했습니다.");
+        return ApiResponse.ok(ResponseCode.POST_UPDATE_SUCCESS);
     }
 
     // 삭제
     @DeleteMapping("/{postId}")
     public ApiResponse<String> updatePost(
-            @RequestHeader("Authorization") String authHeader,
+            @RequestHeader(value = "Authorization", required = false) String authorization,
             @PathVariable("postId") int postId){
 
-        int userId = tokenService.extractUserIdFromHeader(authHeader);
+        if (authorization == null || authorization.isEmpty()){
+            throw new CustomException(ErrorCode.INVALID_TOKEN);
+        }
+
+        int userId = tokenService.extractUserIdFromHeader(authorization);
         postsService.deletePost(userId, postId);
 
-        return ApiResponse.ok("삭제 성공 했습니다.");
+        return ApiResponse.ok(ResponseCode.POST_DELETE_SUCCESS);
     }
 }
